マウスのドラッグした長さで物体を回転させるようにしたい

+++ gravityHit/gravityHit/objChar.h
@@ -22,7 +22,7 @@ public:
 	int ImgHeight{ 0 };//画像の縦サイズ
 
 	int radius{ -1 };
-	int radian{ -1 };
+	float radian{ -1 };
 	Vector VW{ 0,0 }, VH{ 0,0 };
 
 	int Dir{ 0 };//オブジェクトの向き
